requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
Django==3.1
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
djoser==2.0.3
kombu==4.6.11
psycopg2==2.8.5
PyJWT==1.7.1
pytz==2020.1
redis==3.5.3
sqlparse==0.3.1
vine==1.3.0
